Worked on bug :Application Crash while Invite and Share Contact Info

+++ app/src/main/java/com/rawalinfocom/rcontact/adapters/PhoneBookContactListAdapter.java
@@ -93,11 +93,11 @@ public class PhoneBookContactListAdapter extends RecyclerView.Adapter<PhoneBookC
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
                 if (isChecked) {
-                    if (!arrayListCheckedPositions.contains(buttonView.getTag())) {
+                    if (!arrayListCheckedPositions.contains((Integer) buttonView.getTag())) {
                         arrayListCheckedPositions.add((Integer) buttonView.getTag());
                     }
                 } else {
-                    if (arrayListCheckedPositions.contains(buttonView.getTag())) {
+                    if (arrayListCheckedPositions.contains((Integer) buttonView.getTag())) {
                         arrayListCheckedPositions.remove((Integer) buttonView.getTag());
                     }
                 }
